Move DMA buffers to a different file
